==> 5.geoenrichment.ipynb <==
import os
from pathlib2 import Path
home = Path.cwd()
dest = Path(str(home.parents[1]) + '\\gis')
os.chdir(dest)
import esri
from arcgis.features import FeatureLayer
import arcgis.geoenrichment
from arcgis.geoenrichment import enrich
os.chdir(home)
import pandas as pd
import json

def countries(cntry = None):
    '''Returns all countries if no args. Returns country if country arg.'''
    return arcgis.geoenrichment.Country.get(cntry) if cntry else arcgis.geoenrichment.get_countries()

def cbsa(sg, sa = 'metropolitan'):
    '''Returns dict of metropolitan or micropolitan for statistical area chosen for specified subgeography
    and specified statistical area. Default is metropolitan.'''
    return {k: v for (k, v) in sg.subgeographies.cbsa.items() if sa in k.lower()}

def area_ids(d):
    '''Returns a dict of area_id codes to named areas. User enters dict of named areas.'''
    codes_nas = {}
    for k, v in d.items():
        vals = str(v).split()
        for x in vals:
            if 'area_id' in x:
                aid, code = x.split('=')
                #Removing leading/trailing punctuation.
                code = code[1:-2]
                codes_nas[code] = v
    return codes_nas

def credit_calc(atts, dps):
    '''Calculates number of credits used in ArcGIS online for geoenrichment. User enters number of attributes
    that need to be enriched and the number geoenrichment data points.'''
    return (atts / 1000) * 10 * dps

#Have to login to access geoenrichment.
l = esri.Arcgis(str(dest) + '\\config.json')

usa = countries('US')
metros = cbsa(usa)
data = usa.data_collections

#Codes and named areas since cbsa id codes are embedded in named area and don't know how to derive property.
na_codes = area_ids(metros)

#Reading in geolocated records with metro code information in them.
file_dir = '\\'.join(str(dest).split('\\')[0:-1] + ['levels'])
os.chdir(file_dir)
geoc = pd.read_csv('10_11_20_geoc_metros.csv')

#Removing values without a cbsa code.
geoc = geoc[~geoc.loc[:, 'CBSA Code'].isnull()]
#Changing datatype.
geoc.loc[:, 'CBSA Code'] = geoc.loc[:, 'CBSA Code'].astype('int64').astype(str)

#Removing any named areas in the na_codes dict that do not exist in my df.
na_codes = {k: v for (k, v) in na_codes.items() if k in geoc.loc[:, 'CBSA Code'].unique()}

#Seems like 2 codes are not in the arcgis named areas.
len(na_codes) #93
len(geoc.loc[:, 'CBSA Code'].unique()) #95
#2 records for Puerto Rico. Must have been entered in with CBSA codes after the last update.
geoc[~geoc.loc[:, 'CBSA Code'].isin(na_codes)]
#Removing those two records from the dataset.
geoc = geoc[~geoc.index.isin(geoc[~geoc.loc[:, 'CBSA Code'].isin(na_codes)].index)]

#Reading in variables to fetch from ArcGIS content.
with open(Path.joinpath(home, 'variables.json')) as f:
    ag_vars = json.load(f)

#Study areas.
study_ars = list(na_codes.values())

# Enriching data and writing data in parts to avoid accidental credit overusage.

path = Path.joinpath(Path.cwd(), 'geoenrich', 'enriched')

# edu = enrich(study_areas = study_ars, analysis_variables = ag_vars['educ'])
edu.to_csv(Path.joinpath(path, 'educ.csv'), index = False)

# diverse = enrich(study_areas = study_ars, analysis_variables = ag_vars['div'])
diverse.to_csv(Path.joinpath(path, 'diversity.csv'), index = False)

# unemp = enrich(study_areas = study_ars, analysis_variables = ag_vars['unem'])
unemp.to_csv(Path.joinpath(path, 'unemployment.csv'), index = False)

# naics = enrich(study_areas = study_ars, analysis_variables = ag_vars['bus'])
naics.to_csv(Path.joinpath(path, 'businesses.csv'), index = False)

# ent = enrich(study_areas = study_ars, analysis_variables = ag_vars['ent'])
ent.to_csv(Path.joinpath(path, 'entertainment.csv'), index = False)

# crime = enrich(study_areas = study_ars, analysis_variables = ag_vars['crm'])
crime.to_csv(Path.joinpath(path, 'crime.csv'), index = False)

# gens = enrich(study_ars, analysis_variables = ag_vars['gens'])
gens.to_csv(Path.joinpath(path, 'generations.csv'), index = False)

# house = enrich(study_areas = study_ars, analysis_variables = ag_vars['house'])
house.to_csv(Path.joinpath(path, 'housing.csv'), index = False)

# occ = enrich(study_areas = study_ars, analysis_variables = ag_vars['occ'])
occ.to_csv(Path.joinpath(path, 'occupations.csv'), index = False)

# Reading geoenriched information back in to normalize.

dfs = {f.split('.')[0]: pd.read_csv(Path.joinpath(path, f)) for f in os.listdir(Path.joinpath(path)) if f.endswith('.csv')}

#Evaluating unique records.

# for df in dfs.values():
#     print(df.HasData.unique()) #All have ones, indicating data is there for all data.
#     print(df.sourceCountry.unique()) #All US
#     print(df.aggregationMethod.unique()) #All CBSA
#     print(df.populationToPolygonSizeRating.unique()) #All 2.191. Dont know what that means.
#     print(len(df)) #All 93 rows.
#     print(df.columns)

succinct = {k: pd.concat([v.iloc[0:, 3:5], v.iloc[0:, v.columns.get_loc('HasData') + 1:]], axis = 1) for (k, v) in dfs.items()}

#Can actually get rid of shape too...keeping for only one df since geometries are the same.
from functools import reduce
#Beginning result column names.
c_beg = ['StdGeographyID', 'StdGeographyName']
rollup = reduce(lambda df1, df2 : pd.merge(df1, df2, on = c_beg[0]), succinct.values())
#Getting rid of redundant columns.
rollup = rollup.drop([f'{c_beg[1]}_x', f'{c_beg[1]}_y', 'SHAPE_x', 'SHAPE_y'], axis = 1)
#Reordering columns.
cols = c_beg + [col for col in rollup.columns if col not in c_beg]
rollup = rollup[cols]

rollup.loc[:, 'StdGeographyID'] = rollup.loc[:, 'StdGeographyID'].astype(str)

# rollup.to_csv(Path.joinpath(path, 'all.csv'), index = False)

to_visualize = pd.merge(geoc, rollup, left_on = 'CBSA Code', right_on = 'StdGeographyID')

cbsa_url = 'https://services.arcgis.com/P3ePLMYs2RVChkJx/arcgis/rest/services/usa_cbsa/FeatureServer/0'
cbsa_fl = FeatureLayer(cbsa_url)

cbsa_features = cbsa_fl.query(where = "CBSA_TYPE = 'Metropolitan'")
#Creating spatial dataframe.
cbsa_df = cbsa_features.sdf

#Segmenting out poughkeepsie prior to join since code not represented in CBSA metro df.
poughkeepsie = to_visualize[to_visualize.loc[:, 'CBSA Code'] == '39100']
to_visualize = to_visualize[to_visualize.loc[:, 'CBSA Code'] != '39100']

poughkeepsie.loc[:, 'CBSA Code'] = '35620'

#Bringing back in to original df.
to_visualize = pd.concat([to_visualize, poughkeepsie], axis = 0)

to_visualize = pd.merge(to_visualize, cbsa_df[['CBSA_ID', 'SQMI']], left_on = 'CBSA Code', right_on = 'CBSA_ID')

#Identifying coolumns that need to be divided by square mile.
vars_psqm = list({k: v for (k, v) in ag_vars.items() if k not in ['crm', 'ent', 'pop', 'disp', 'ind_emp']}.values())
vars_psqm = [i for l in vars_psqm for i in l]
#Removing index items that were embedded in original lists.
vars_psqm.remove('DIVINDX_CY')
vars_psqm.remove('HAI_CY')

#Creating per square mile calculated columns for appropriate columns.
for c in vars_psqm:
    to_visualize.loc[:, c + '_sqmi'] = to_visualize.loc[:, c] / to_visualize.loc[:, 'SQMI']

#Dropping SHAPE (geometry) column for now for ease of viewing tabular data.
to_visualize = to_visualize.drop('SHAPE', axis = 1)

to_visualize.to_csv('10_11_20_enriched.csv', index = False)
